chore: rename nova api calls, fix namespace for NovaTestTrait
